don't pass agenda_args to parent class
The starting point of this commit is to make MySQLAgenda consistent in the
way it accesses the db_name. Before this commit it was from its field in
`createDB` but from the `$agenda_args` passed as a parameter from `openDB`.

While refactoring this I figured that since `$agenda_args` is an arbitrary array
with keys which depends on the implementation, it makes sense to leave it at
the level of the implementations.