Add test_addAndRemoveEvent
This commit also adds a MockCalDAVClient because it is not convenient
to have this kind of dynamic behavior with phpunit mock.
The existing tests now use this new mock in order to avoid duplication
in the code which implements the mocked method (except for tests that had
simpler mock and which relied on expectation on the number of calls).

This also fixes a bug we had in the previous mock for $eventsIdToEtag
(it's not supposed to be an array of array, but a key-value pair array).
This bug wasn't an issue as long as the agenda isn't reloaded so it didn't
impacted the existing tests.