Removed unused and buggy Agenda::saveEvent
I noticed that this method is unused because the caldav server we use in
prod doesn't return an ETag. We can see it for instance in the logs.

It also appears that this method is buggy (that's because we don't use it
that we didn't notice it). In particular:
- we call it with $vCalendar->serialize() but the implementation declares
  that it expects an object
- the PDO query declares placeholder ":vCalendarRaw" but then we define
  "vCalendar"
- this does not update the list of attendees

We could probably fix all this, but OTHO, since we don't use it anyway,
it seems easier to delete it (at least for now: it will always be possible
to reintroduce this feature in the future if the need arises)