All registrations may create a reminder
This fixes #100.
The issue was that when a user register to an event which is displayed on
Slack after calling /remindEvent, then the registration is runs
`SlackEvents.more_inchannel` which doesn't call `Agenda.addReminder`.

This commit moves the creation of the reminder inside `Agenda.updateAttendee`
so every registration leads to the attempt to create a reminder.