Make "now" consistently injected
Before this commit:
- part of Agenda relied on some injected value for $now (which is
  needed to have deterministic tests)
- and part of it computed some "now" based on the system clock

This made the tests running in an inconsistent state after the previous
commit because addReminder (now called by updateAttendee) compared the
star date of events (based on injected $now) with the current time
(based on system clock).

This commit fixes it by relying on the injected $now everywhere.
(Rather than passing $now around on most method of this object, it was
 more relevant to make it a class field)

However this commit couldn't do just this because this refacto turned
some tests red. That's because now that we're in a consistent state,
when we run addReminder we now try to call the slack API to register
a reminder. We hence need to configure the mock when it is called on this
method. We use this opportunity to also add expectations on the number
of calls of this method.

Note that $now is injected we could go one step farther and add expectations
on the number of call to the slackAPI mock in the tests about user removing
themselves from events, but that will be for a next commit.