Agenda.update does a bulk select
Doing a single query is a bit better for performances.

I just tested this patch in preprod where we currently have
~30 events on the caldav server.
When I delete an event, this method runs in about 8ms without this patch
and it only takes 4ms with it.

It's admittedly not a lot since the whole query take ~0.9s, but that's
still a gain

(this relates to #63)