Limit the number of events returned
Because slack has a limit on the number of items we can return.
It turns out that we reach this limit when we return more than 30
events.
Ideally we should add some pagination to adress this, but for now
a hard limit will avoid this issue.
(it will prevent users to see events farter in the future, but they
wouldn't be able to see it even without this limit since the app
would crash. At least with this patch they can still see the first events
and they can use category filters to see events farther in the future)