Be robust in case of non UTF-8 characters
Having non UTF-8 characters is a case we may encounter if a user creates an
event with a bad encoding. And by default json_encode returns "FALSE" if it
encounters non UTF-8 characters. In such case, the call to the slack API
subsequently fails because the payload is incorrect (it has "FALSE" instead
of an expected json payload).

Adding this flag seems to make is possible to cope correctly with such bad
input