Don't log a 412 as an error when it is expected
Since the caldav server may be edited concurrently by several clients,
it is expected that we may get a 412 when we first try to update an event
(and we handle this case with a retry after a re-sync).
In this case it is pointless to log an error because we don't need to worry
about it.
With this patch such situation is now logged as an info.

This fixes #141