Upgrade package to enhance compatibility with PHP 8.1
currently our prod is in PHP 8.0 but when we tried updating to 8.1 we
observed some errors with a stacktrace in the middle of the sabre/vobject
code.

We require a more recent version in composer.json in order to
fix this compatibility.
In the process we ran a "composer update" which update all it could
in composer.lock