Add some log in order identify our bottlenecks
It's not so rare that a query takes at least 1s to be handled.
It's rather slow in itself, and it may be an issue since Slack has a 3s
timeout (that we may fail to respect some times).

There may be ways to speed up our code (eg: caching on the filesystem
some complex sql queries for which we would have the certainty that
the data would remain up to date).

This commit adds some logs, and since each log it dumped with a timestamp,
this should make it easier to understand where the time is going and hence
where it could make sense to think of some caching system (and subsequently,
if we end up developing some optimizations, it could help checking that
said optimizations would actually save time)